lineBot_Server/requirements.txt
asgiref==3.4.1
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
Django==3.2.7
Flask==2.0.1
future==0.18.2
idna==3.2
isort==5.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.4.3
line-bot-sdk==1.20.0
MarkupSafe==2.0.1
mccabe==0.6.1
pylint==2.6.0
pytz==2021.1
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
gunicorn==19.4.5